Fix Kotlin compilation and type errors in widget and opoc util packages
This commit resolves over 50 pre-existing syntax and type errors from an incomplete migration to Kotlin in the `widget` and `net.gsantner.opoc.util` packages. Key changes include:
- Fixed `FastAdapter` breaking changes (v5 migration from FastItemAdapter to ItemAdapter).
- Fixed Kotlin `nullability` errors (using safe calls `?.` and explicit assertions `!!`).
- Resolved unresolved symbols and missing methods in `ContextUtils.kt`, `ShareUtil.kt`, `NetworkUtils.kt` and `FileUtils.kt`.
- Added missing `SimpleFingerGestures` dependency to `build.gradle` and fixed its imports (`in` escaped).
- Fixed numeric property type mismatches requiring `.toFloat()` conversions.

+++ app/src/main/kotlin/com/jiadanni/launcher4d/widget/AppDrawerPage.kt
@@ -132,7 +132,7 @@ class AppDrawerPage @JvmOverloads constructor(
                     cardView.cardElevation = 0f
                 } else {
                     cardView.setCardBackgroundColor(Setup.appSettings().drawerCardColor)
-                    cardView.cardElevation = Tool.dp2px(4)
+                    cardView.cardElevation = Tool.dp2px(4).toFloat()
                 }
 
                 val cc = layout.findViewById<CellContainer>(R.id.group)
